+++ include/geneva/GSerialGD.hpp
@@ -83,7 +83,7 @@ public:
 	/** @brief The default constructor */
 	GSerialGD();
 	/** @brief Initialization with the number of starting points and the size of the finite step */
-	GSerialGD(const std::size_t&, const float&, const float&);
+	GSerialGD(const std::size_t&, const double&, const double&);
 	/** @brief A standard copy constructor */
 	GSerialGD(const GSerialGD&);
 	/** @brief The destructor */
@@ -99,12 +99,12 @@ public:
 
 	/** @brief Checks whether this object fulfills a given expectation in relation to another object */
 	virtual boost::optional<std::string> checkRelationshipWith(
-			const GObject&
-			, const Gem::Common::expectation&
-			, const double&
-			, const std::string&
-			, const std::string&
-			, const bool&
+      const GObject&
+      , const Gem::Common::expectation&
+      , const double&
+      , const std::string&
+      , const std::string&
+      , const bool&
 	) const OVERRIDE;
 
 	/** @brief Adds local configuration options to a GParserBuilder object */
